Fix code

+++ src/services/plant.js
@@ -26,9 +26,6 @@ export default class PlantService {
     return plant
   }
 
-  /* TODO : Checker pourquoi le traitement commenter ne fonctionne pas
-   const plant = await Plant.update({ ...fields }, { where: { plant_id: id } })
-   TODO : Le code si dessous ne gère pas les appostrophe lors d'un update dans une chaine ( TODO FRONT ) */
   async putPlantById ({ fields, id, userId }) {
     const plant = await Plant.findByPk(id)
     if (!plant) {
@@ -40,15 +37,16 @@ export default class PlantService {
     await plant.update(fields)
     await plant.save()
     return plant
-    // const plant = await Plant.update({ ...fields }, { where: { plant_id: id } })
-    // const plant = await sequelize.query(`UPDATE PLANT SET name_plant='${fields.name_plant}', instructions_plant='${fields.instructions_plant}', type_plant='${fields.type_plant}',status_plant='${fields.status_plant}' WHERE plant_id = ${id}`)
-    // if (!plant) {
-    //   throw new NotFound('La plant n\'existe pas.')
-    // }
-    // return plant
   }
 
-  async deletePlantById ({ id }) {
+  async deletePlantById ({ id, userId }) {
+    const plant = await Plant.findByPk(id)
+    if (!plant) {
+      throw new NotFound('La plante n\'existe pas.')
+    }
+    if (plant.user_id !== userId) {
+      throw new Unauthorized('Vous n\'êtes pas le propriétaire de cette plante.')
+    }
     await Plant.destroy({ where: { plant_id: id } })
   }
